Added new files, and compile by default now single and double prec version.

+++ main/microhh_single.cxx
@@ -37,7 +37,7 @@ int main(int argc, char *argv[])
         master.print_message("Microhh git-hash: " GITHASH "\n");
 
         // Initialize the model class.
-        Model<double> model(&master, argc, argv);
+        Model<float> model(&master, argc, argv);
 
         // Initialize the model components.
         model.init();
